Fix linter issues

+++ morpheus/framework/resources/network_router_firewall_rule/resource.go
@@ -111,7 +111,7 @@ func (r *Resource) Create(
 		return
 	}
 
-	id := int64(*result.Id.Get())
+	id := *result.Id.Get()
 	plan.Id = types.Int64Value(id)
 
 	state, pdiags := getFirewallRuleAsState(ctx, id, routerID, client, plan)
