+++ morpheus/framework/resources/network_router_route/resource.go
@@ -127,7 +127,7 @@ func (r *Resource) Create(
 		return
 	}
 
-	id := int64(*result.Id.Get())
+	id := *result.Id.Get()
 	plan.Id = types.Int64Value(id)
 
 	state, pdiags := getRouteAsState(ctx, id, routerID, client, plan)
